kracken/requirements.txt
# Core dependencies
ccxtpro>=4.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
numpy>=1.21.0
pandas>=1.3.0
psycopg2-binary>=2.9.3
python-telegram-bot>=13.7
requests>=2.26.0

# Machine Learning
tensorflow>=2.8.0
scikit-learn>=1.0.0

# Quantum Computing (optional)
qiskit>=0.34.0
qiskit-finance>=0.3.0

# Technical Analysis
TA-Lib>=0.4.24

# Backtesting
backtesting>=0.3.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2021.3

